chore: make directory for each role
패키지 구조를 모델, 레포지토리, 뷰로 나누어 해당하는 파일들을 분류
