--- app/src/main/java/campus/tech/kakao/contacts/Gender.kt
@@ -1,5 +1,0 @@
-package campus.tech.kakao.contacts
-
-enum class Gender {
-    FEMALE,MALE
-}
