+++ app/src/main/java/campus/tech/kakao/contacts/Model/Gender.kt
@@ -1,0 +1,5 @@
+package campus.tech.kakao.contacts.Model
+
+enum class Gender {
+    FEMALE,MALE
+}
